all: fix every warning

+++ src/drivers/hollywood_sdmmc.c
@@ -36,7 +36,7 @@ static ssize_t sdmmcRead(struct blockDevice *bdev, void *dest, size_t len, u64 o
 	blkSize = bdev->blockSize;
 	startBlock = (size_t)(off / blkSize);
 	remaining = len;
-	headSkip = off % blkSize;
+	headSkip = (size_t)(off % blkSize);
 	headBytes = alignedBlocks = tailBytes = 0;
 
 	//log_printf("read: dest=0x%08x, SB=%u, len=%u, HS=%u, HB=%u\r\n", dest, startBlock, len, headSkip, headBytes);
@@ -50,7 +50,7 @@ static ssize_t sdmmcRead(struct blockDevice *bdev, void *dest, size_t len, u64 o
 	/* if we have an unaligned first block, read it separately */
 	if (headSkip) {
 		dcache_invalidate(tmp, sizeof(tmp));
-		ret = mmc_block_read(mmcDev, startBlock, 1, tmp,
+		ret = (size_t)mmc_block_read(mmcDev, startBlock, 1, tmp,
 			(uintptr_t)virtToPhys(tmp), NULL, NULL);
 		if (ret != blkSize)
 			return -1;
@@ -72,7 +72,7 @@ static ssize_t sdmmcRead(struct blockDevice *bdev, void *dest, size_t len, u64 o
 	/* read the remaining aligned portions in one pass, if any */
 	if (alignedBlocks) {
 		dcache_invalidate(dest, alignedBlocks * blkSize);
-		ret = mmc_block_read(mmcDev, startBlock, alignedBlocks, dest,
+		ret = (size_t)mmc_block_read(mmcDev, startBlock, alignedBlocks, dest,
 			(uintptr_t)virtToPhys(dest), NULL, NULL);
 		if (ret != alignedBlocks * blkSize)
 			return -1;
@@ -84,7 +84,7 @@ static ssize_t sdmmcRead(struct blockDevice *bdev, void *dest, size_t len, u64 o
 	/* if we have an unaligned last block, read it separately */
 	if (tailBytes) {
 		dcache_invalidate(tmp, sizeof(tmp));
-		ret = mmc_block_read(mmcDev, startBlock, 1, tmp,
+		ret = (size_t)mmc_block_read(mmcDev, startBlock, 1, tmp,
 			(uintptr_t)virtToPhys(tmp), NULL, NULL);
 		if (ret != blkSize)
 			return -1;
@@ -92,7 +92,7 @@ static ssize_t sdmmcRead(struct blockDevice *bdev, void *dest, size_t len, u64 o
 		memcpy(dest, tmp, tailBytes);
 	}
 
-	return len;
+	return (ssize_t)len;
 }
 
 static ssize_t sdmmcWrite(struct blockDevice *bdev, const void *src, size_t len, u64 off) {
